+++ controlplane/src/types/index.ts
@@ -1,7 +1,14 @@
-import { LintSeverity } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
+import {
+  CompositionError,
+  CompositionWarning,
+  DeploymentError,
+  LintSeverity,
+} from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
 import { JWTPayload } from 'jose';
+import { PlainMessage } from '@bufbuild/protobuf';
 import { DBSubgraphType, GraphPruningRuleEnum, OrganizationRole, ProposalMatch, ProposalOrigin } from '../db/models.js';
 import { RBACEvaluator } from '../core/services/RBACEvaluator.js';
+import { ComposeGraphsTaskResultItem } from '../core/composition/composeGraphs.types.js';
 
 export const COMPOSITION_IGNORE_EXTERNAL_KEYS_FEATURE_ID = 'composition-ignore-external-keys';
 export const SPLIT_CONFIG_LOADING_FEATURE_ID = 'split-config-loading';
@@ -806,4 +813,20 @@ export interface ProposalSubgraphDTO {
   currentSchemaVersionId?: string;
   isNew: boolean;
   labels: Label[];
+}
+
+export interface ComposeAndDeployResult {
+  deploymentErrors: PlainMessage<DeploymentError>[];
+  compositionErrors: PlainMessage<CompositionError>[];
+  compositionWarnings: PlainMessage<CompositionWarning>[];
+}
+
+export interface OrganizationFeatures {
+  ignoreExternalKeys: boolean;
+  splitConfigLoading: boolean;
+}
+
+export interface FederatedGraphAndCompositionResults {
+  federatedGraph: FederatedGraphDTO;
+  results: ComposeGraphsTaskResultItem[];
 }
